Updated the confirmation phase :
- Added conclusion with `ConversationHandler.END` : sends a nice sticker check it out
- Added a sticker handling exception. that's simply for developper to conveniently find sticker id. Should be put in another command maybe.
- Modified what was needed to be able to chang easily name of the meal in the confirmation phase. Still need to implement that for the rest of the items.

+++ foodshare/handlers/cook_conversation/meal_name_choice.py
@@ -3,11 +3,14 @@ from telegram import ChatAction
 from foodshare.commands.gif_test import get_gif_url
 
 from . import ConversationStage
+from .conclusion_selection import ask_for_conclusion
 from .date_selection import ask_for_date
 
 
 def ask_for_meal_name(update, context):
-    context.user_data.clear()
+    ud = context.user_data
+    if 'confirmation_stage' not in ud:
+        ud.clear()
     message = (
         'Tell me what you want to cook! '
         '(just type the name of the meal as an answer to this message)'
@@ -18,7 +21,6 @@ def ask_for_meal_name(update, context):
         update.callback_query.edit_message_text(text=message,)
     else:
         update.message.reply_text(message)
-
     return ConversationStage.TYPING_MEAL_NAME
 
 
@@ -26,9 +28,10 @@ def save_meal_name(update, context):
     ud = context.user_data
     bot = context.bot
     chat_id = update.message.chat_id
-
+    ud['chat_id'] = chat_id
     ud['meal_name'] = update.message.text
-
+    if 'confirmation_stage' in ud:
+        return ask_for_conclusion(update, context, highlight='meal_name')
     bot.send_chat_action(chat_id=chat_id, action=ChatAction.TYPING)
     bot.deleteMessage(chat_id=chat_id, message_id=update.message.message_id)
 
@@ -39,3 +42,13 @@ def save_meal_name(update, context):
         update.message.reply_text(f'No gif found for {ud["meal_name"]}')
 
     return ask_for_date(update, context)
+
+
+def return_sticker(
+    update, context
+):  # Only for developpment to know sticker id
+    file_id = update.message.sticker['file_id']
+    bot = context.bot
+    chat_id = update.message.chat_id
+    print(file_id)
+    bot.send_sticker(chat_id, file_id)
